requirements.txt
django==3.2
django-debug-toolbar==3.2
djangorestframework==3.12.4
model_bakery==1.2.1
